backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
sqlalchemy
psycopg2-binary
requests
feedparser
google-generativeai
